requirements.txt
openai>=1.0.0
chromadb>=0.4.0
pypdf>=3.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
streamlit>=1.30.0
